added initial config and bot logic for upcoming mock testing framework, map server communications to std(in|out) and logs to stderr

+++ ircbot.sh
@@ -175,22 +175,33 @@ if [ -z "$SERVER" ]; then
     die "A server must be defined; check the configuration."
 fi
 
-# Connect to server
-[ -n "$TLS" ] && TLS="--ssl"
-if [ -z "$BASH_TCP" ]; then
-    exec 3<> "$infile" ||
-        die "unknown failure mapping named pipe ($infile) to fd"
-    exec 4<> "$outfile" ||
-        die "unknown failure mapping named pipe ($outfile) to fd"
-    ( ncat "$SERVER" "${PORT:-6667}" "$TLS" <&3 >&4
-      echo "*** CRITICAL *** ncat closed unexpectedly" >&2
-      kill -USR1 $$ ) &
+# this mode should be used for testing only
+if [ -z "$MOCK_CONN_TEST" ]; then
+    # Connect to server
+    [ -n "$TLS" ] && TLS="--ssl"
+    if [ -z "$BASH_TCP" ]; then
+        exec 3<> "$infile" ||
+            die "unknown failure mapping named pipe ($infile) to fd"
+        exec 4<> "$outfile" ||
+            die "unknown failure mapping named pipe ($outfile) to fd"
+        ( ncat "$SERVER" "${PORT:-6667}" "$TLS" <&3 >&4
+          echo "*** CRITICAL *** ncat closed unexpectedly" >&2
+          kill -USR1 $$ ) &
+    else
+        infile="/dev/tcp/${SERVER}/${PORT}"
+        exec 3<> "$infile" ||
+            die "Cannot connect to ($SERVER) on port ($PORT)"
+        exec 4<&3 ||
+            die "unknown failure mapping named pipe ($infile) to fd"
+    fi
 else
-    infile="/dev/tcp/${SERVER}/${PORT}"
-    exec 3<> "$infile" ||
-        die "Cannot connect to ($SERVER) on port ($PORT)"
-    exec 4<&3 ||
-        die "unknown failure mapping named pipe ($infile) to fd"
+    # send irc communication to
+    exec 4>&0 # from server stdin
+    exec 3<&1 # to   server stdout
+    exec 1>&-
+    exec 1<&2 # remap stdout to err for logs
+    # full logging
+    LOG_LEVEL=1
 fi
 
 #################
